infrastructure/lambda/orchestration/requirements.txt
boto3>=1.28.0
botocore>=1.31.0
psycopg[binary]>=3.1.0
